refactor: move stations.dart from application to presentation
