+++ bridge_simulator/lib/bridge_simulator.dart
@@ -1,5 +1,5 @@
 /// A bridge simulator for Star Trek's Galaxy-class starships.
 library;
 
-export 'src/application/stations.dart';
+export 'src/presentation/stations.dart';
 export 'src/presentation/buttons.dart';
